Day 3 part 2

+++ 2024/src/day3/Part1.java
@@ -3,8 +3,6 @@ package day3;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -44,7 +42,7 @@ public class Part1 {
       final int secondNumber = Integer.parseInt(numberMatcher.group());
       //System.out.println("First number: " + firstNumber);
       //System.out.println("Second number: " + secondNumber);
-      total += firstNumber * secondNumber;
+      total += (long)firstNumber * secondNumber;
     }
 
     System.out.println("Total: " + total);
